Revert "Integrate MoneyRequestConfirmationListFooter"
